requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
groq==0.13.1
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.9.2
pandas==2.2.2
streamlit==1.37.1
